requirements.txt
Django==1.7
Pillow==2.5.3
dj-database-url==0.3.0
dj-static==0.0.6
django-friendship==1.1.0
django-timedeltafield==0.7.3
django-toolbelt==0.0.1
gunicorn==19.0.0
psycopg2==2.5.3
static3==0.5.1
wsgiref==0.1.2
